+++ src/lib/input-validation/constant/errorType.ts
@@ -15,5 +15,5 @@ export const CommonValidationErrorType = Enum(
   "invalidFormat",
   "lessThanMinimumValue",
   "exceedMaximumValue",
-  "notANumber"
+  "invalidValue"
 );